backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
websockets==12.0
pandas==2.2.0
numpy==1.26.0
anthropic==0.39.0
ccxt==4.4.0
python-dotenv==1.0.1
